refactor ask service and tighten request files
